TVGUI-10 Moving EditReminder into a component
